+++ game_libs/common/entity_state.h
@@ -147,7 +147,7 @@ typedef struct clientdata_s
 	float		fov;
 	int		weaponanim;
 	int		weaponScreenOverlay;
-	byte	weaponInaccuracy;
+	float	weaponInaccuracy;
 
 	int		m_iId;
 	float		m_flNextAttack;
